Storage Path is added & File,Folder Db is created

+++ src/main/java/com/example/swaraj/SkyLock/Repo/UsersRepo.java
@@ -4,8 +4,11 @@ import com.example.swaraj.SkyLock.Models.Users;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
-public interface UsersRepo extends JpaRepository<Users,Integer> {
+public interface UsersRepo extends JpaRepository<Users, Long> {
 
     Users findByUsername(String username);
+
 }
